Improvement: put two tabs for file and text submission into one tab
Related to #212
